+++ src/main/java/saigonwithlove/ivy/intellij/shared/Configuration.java
@@ -1,16 +1,21 @@
 package saigonwithlove.ivy.intellij.shared;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 
-@Getter
+@Data
 @Builder
+@NoArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class Configuration {
   private String name;
-  @Setter private String value;
+  private String value;
   private String defaultValue;
+  private String description;
 
   public boolean isModified() {
     return value != null && !StringUtils.equals(value, defaultValue);