+++ src/main/java/saigonwithlove/ivy/intellij/shared/Configurations.java
@@ -2,6 +2,7 @@ package saigonwithlove.ivy.intellij.shared;
 
 import com.google.common.base.Preconditions;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.util.xmlb.Converter;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -13,6 +14,7 @@ import java.util.stream.Collectors;
 import lombok.experimental.UtilityClass;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 @UtilityClass
 public class Configurations {
@@ -68,17 +70,31 @@ public class Configurations {
   }
 
   @NotNull
-  public List<Configuration> buildConfigurations(
-      Map<String, String> defaultConfigurations, Map<String, String> modifiedConfigurations) {
-    return defaultConfigurations.entrySet().stream()
+  public List<Configuration> buildConfigurations(Map<String, Configuration> configurations) {
+    return configurations.entrySet().stream()
         .map(
             entry ->
                 Configuration.builder()
                     .name(entry.getKey())
-                    .defaultValue(entry.getValue())
-                    .value(modifiedConfigurations.get(entry.getKey()))
+                    .defaultValue(entry.getValue().getDefaultValue())
+                    .value(entry.getValue().getValue())
                     .build())
         .sorted((a, b) -> a.getName().compareToIgnoreCase(b.getName()))
         .collect(Collectors.toList());
+  }
+
+  public static class StateConverter extends Converter<Configuration> {
+
+    @Nullable
+    @Override
+    public Configuration fromString(@NotNull String value) {
+      return null;
+    }
+
+    @Nullable
+    @Override
+    public String toString(@NotNull Configuration value) {
+      return null;
+    }
   }
 }