+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.2.0 - Unreleased]
+### Planned
+- Support Ivy 8.
+
+### Changed
+- Use RxJava to manage State change.
+- Implement the new way to manage Ivy Engine, the handling for each Ivy version is separated into different classes.
+
 ### Fixed
-- Cannot restore Server Properties in State
+- Cannot restore Server Properties in State.
 
 ## [v0.1.12] - 2020-05-03
 ### Added